remove redundant static_cast

+++ projects/composablekernel/include/ck_tile/core/arch/mma/scale/mfma/scale_gfx9.hpp
@@ -10,6 +10,7 @@
 #include "ck_tile/core/arch/mma/scale/scale_traits.hpp"
 #include "ck_tile/core/config.hpp"
 #include "ck_tile/core/numeric/float8.hpp"
+#include "ck_tile/core/numeric/integer.hpp"
 #include "ck_tile/core/numeric/pk_f6.hpp"
 #include "ck_tile/core/numeric/pk_fp4.hpp"
 #include "ck_tile/core/numeric/vector_type.hpp"
@@ -40,8 +41,11 @@ struct amdgcn_mma<fp8_t, fp8_t, fp32_t, 16u, 16u, 128u, CtrlFlags, CompilerTarge
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4(
             bit_cast<int32x8_t>(aVec),
@@ -49,9 +53,9 @@ struct amdgcn_mma<fp8_t, fp8_t, fp32_t, 16u, 16u, 128u, CtrlFlags, CompilerTarge
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<fp8_t>,
             scale::detail::ScaleDataTypeToFlag_v<fp8_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -79,8 +83,11 @@ struct amdgcn_mma<bf8_t, bf8_t, fp32_t, 16u, 16u, 128u, CtrlFlags, CompilerTarge
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4(
             bit_cast<int32x8_t>(aVec),
@@ -88,9 +95,9 @@ struct amdgcn_mma<bf8_t, bf8_t, fp32_t, 16u, 16u, 128u, CtrlFlags, CompilerTarge
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<bf8_t>,
             scale::detail::ScaleDataTypeToFlag_v<bf8_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -118,8 +125,11 @@ struct amdgcn_mma<pk_fp4_t, pk_fp4_t, fp32_t, 16u, 16u, 128u, CtrlFlags, Compile
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         int32x4_t arg_a = bit_cast<int32x4_t>(aVec);
         int32x4_t arg_b = bit_cast<int32x4_t>(bVec);
@@ -130,9 +140,9 @@ struct amdgcn_mma<pk_fp4_t, pk_fp4_t, fp32_t, 16u, 16u, 128u, CtrlFlags, Compile
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp4_t>,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp4_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -159,8 +169,11 @@ struct amdgcn_mma<pk_fp6x16_t, pk_fp6x16_t, fp32_t, 16u, 16u, 128u, CtrlFlags, C
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4(
             int32x8_t{aVec.data[0], aVec.data[1], aVec.data[2], aVec.data[3], aVec.data[4], aVec.data[5], 0, 0},
@@ -168,9 +181,9 @@ struct amdgcn_mma<pk_fp6x16_t, pk_fp6x16_t, fp32_t, 16u, 16u, 128u, CtrlFlags, C
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp6x16_t>,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp6x16_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -197,9 +210,12 @@ struct amdgcn_mma<pk_bf6x16_t, pk_bf6x16_t, fp32_t, 16u, 16u, 128u, CtrlFlags, C
 {
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4";
-        
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_16x16x128_f8f6f4(
             int32x8_t{aVec.data[0], aVec.data[1], aVec.data[2], aVec.data[3], aVec.data[4], aVec.data[5], 0, 0},
@@ -207,9 +223,9 @@ struct amdgcn_mma<pk_bf6x16_t, pk_bf6x16_t, fp32_t, 16u, 16u, 128u, CtrlFlags, C
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<pk_bf6x16_t>,
             scale::detail::ScaleDataTypeToFlag_v<pk_bf6x16_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -238,8 +254,11 @@ struct amdgcn_mma<fp8_t, fp8_t, fp32_t, 32u, 32u, 64u, CtrlFlags, CompilerTarget
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4(
             bit_cast<int32x8_t>(aVec),
@@ -247,9 +266,9 @@ struct amdgcn_mma<fp8_t, fp8_t, fp32_t, 32u, 32u, 64u, CtrlFlags, CompilerTarget
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<fp8_t>,
             scale::detail::ScaleDataTypeToFlag_v<fp8_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -277,8 +296,11 @@ struct amdgcn_mma<bf8_t, bf8_t, fp32_t, 32u, 32u, 64u, CtrlFlags, CompilerTarget
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4(
             bit_cast<int32x8_t>(aVec),
@@ -286,9 +308,9 @@ struct amdgcn_mma<bf8_t, bf8_t, fp32_t, 32u, 32u, 64u, CtrlFlags, CompilerTarget
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<bf8_t>,
             scale::detail::ScaleDataTypeToFlag_v<bf8_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -316,8 +338,11 @@ struct amdgcn_mma<pk_fp4_t, pk_fp4_t, fp32_t, 32u, 32u, 64u, CtrlFlags, Compiler
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         int32x4_t arg_a = bit_cast<int32x4_t>(aVec);
         int32x4_t arg_b = bit_cast<int32x4_t>(bVec);
@@ -328,9 +353,9 @@ struct amdgcn_mma<pk_fp4_t, pk_fp4_t, fp32_t, 32u, 32u, 64u, CtrlFlags, Compiler
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp4_t>,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp4_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -357,8 +382,11 @@ struct amdgcn_mma<pk_fp6x16_t, pk_fp6x16_t, fp32_t, 32u, 32u, 64u, CtrlFlags, Co
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4(
             int32x8_t{aVec.data[0], aVec.data[1], aVec.data[2], aVec.data[3], aVec.data[4], aVec.data[5], 0, 0},
@@ -366,9 +394,9 @@ struct amdgcn_mma<pk_fp6x16_t, pk_fp6x16_t, fp32_t, 32u, 32u, 64u, CtrlFlags, Co
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp6x16_t>,
             scale::detail::ScaleDataTypeToFlag_v<pk_fp6x16_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
@@ -396,8 +424,11 @@ struct amdgcn_mma<pk_bf6x16_t, pk_bf6x16_t, fp32_t, 32u, 32u, 64u, CtrlFlags, Co
     static constexpr const char* instruction_name =
         "__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4";
 
-    CK_TILE_DEVICE static CVecType
-    exec(AVecType const& aVec, BVecType const& bVec, CVecType const& cVec, int scale_A, int scale_B)
+    CK_TILE_DEVICE static CVecType exec(AVecType const& aVec,
+                                        BVecType const& bVec,
+                                        CVecType const& cVec,
+                                        int32_t scale_A,
+                                        int32_t scale_B)
     {
         return {__builtin_amdgcn_mfma_scale_f32_32x32x64_f8f6f4(
             int32x8_t{aVec.data[0], aVec.data[1], aVec.data[2], aVec.data[3], aVec.data[4], aVec.data[5], 0, 0},
@@ -405,9 +436,9 @@ struct amdgcn_mma<pk_bf6x16_t, pk_bf6x16_t, fp32_t, 32u, 32u, 64u, CtrlFlags, Co
             cVec,
             scale::detail::ScaleDataTypeToFlag_v<pk_bf6x16_t>,
             scale::detail::ScaleDataTypeToFlag_v<pk_bf6x16_t>,
-            static_cast<int>(CtrlFlags::OPSEL_A),
+            CtrlFlags::OPSEL_A,
             scale_A,
-            static_cast<int>(CtrlFlags::OPSEL_B),
+            CtrlFlags::OPSEL_B,
             scale_B)};
     }
 };
